Refactor Hyundai CAN-FD: Remove `hyundai_canfd.h`, consolidate CAN-FD utilities into `hyundai_common.h`
- Fully deprecated `hyundai_canfd.h` by removing unused and redundant SCC message definitions, macros, and constants.
- Moved essential utility functions (`hyundai_canfd_get_counter`, `hyundai_canfd_get_checksum`) to `hyundai_common.h` for better modularity and maintainability.
- Updated includes and definitions across modules to reference `hyundai_common.h`, ensuring consistency.

**Safety-Critical Changes:**
- Eliminated unused address check macros for SCC_CONTROL and button configurations to reduce potential for error.
- Improved modularity by centralizing CAN-FD utilities, reducing reliance on scattered definitions.
- No changes to existing safety behavior, ensuring preserved functionality.
